cleaned up code in backend dir and created separate files for validations
