+++ backend/routes/api/bookings.js
@@ -1,60 +1,10 @@
 const express = require('express');
 const router = express.Router();
 
-const { requireAuth, reviewPermission, bookingBelongsPermission, spotOwnerOrBookingOwnerPermission } = require('../../utils/auth');
-const { Spot, User, Review, Image, Booking, sequelize } = require('../../db/models');
-const { check, body } = require('express-validator');
-const { handleValidationErrors } = require('../../utils/validation');
+const { requireAuth, bookingBelongsPermission, spotOwnerOrBookingOwnerPermission } = require('../../utils/auth');
+const { Spot, User, Image, Booking } = require('../../db/models');
 const { Op } = require('sequelize');
-
-const validateBooking = [
-    body('endDate')
-    .custom((value, { req }) => {
-        if (value < req.body.startDate) throw new Error ('endDate cannot come before startDate')
-        return true
-    }),
-    handleValidationErrors
-]
-
-// Middleware that checks if booking exists
-const existsBooking = async (req, res, next) => {
-    const booking = await Booking.findByPk(req.params.bookingId);
-
-    if (!booking) {
-        const err = new Error ("Booking couldn't be found");
-        err.status = 404;
-        return next(err);
-    }
-    return next();
-}
-
-// Can't edit a booking that is in the past aka already happened
-const bookingPastEndDate = async (req, res, next) => {
-
-    const booking = await Booking.findByPk(req.params.bookingId, {
-        attributes: ['endDate']
-    });
-    const today = new Date();
-    // const date = today.getFullYear() + '-' + (today.getMonth()+1) + '-' + today.getDate();
-
-
-    const bookingEndDate = new Date(booking.endDate)
-
-    // console.log('TEST', booking.endDate);
-    // console.log('TODAYS DATE', date);
-    // console.log(new Date());
-    // if (date > booking.dataValues.endDate[0]) {
-    //     const err = new Error ("Past bookings can't be modified");
-    //     err.status = 400;
-    //     return next(err);
-    // }
-    if (today.getTime() > bookingEndDate.getTime()) {
-        const err = new Error ("Past bookings can't be modified");
-        err.status = 400;
-        return next(err);
-    }
-    return next();
-}
+const { validateBooking, existsBooking, bookingPastEndDate, pastStartDate } = require('../../utils/booking-validations')
 
 // Booking conflict middleware (NEED TO REFACTOR THIS LATER MAYBE USE "check" or "body")
 const bookingConflictErr = async (req, res, next) => {
@@ -86,54 +36,13 @@ const bookingConflictErr = async (req, res, next) => {
         }
     })
 
-    // console.log('spot start end', spotStartEnd)
-
     if (spotStartEnd) {
         const err = new Error ('Sorry, this spot is already booked for the specifed dates');
         err.status = 403;
         err.errors = {startDate: 'Start date conflicts with an exiting booking',endDate: 'End date conflicts with an existing booking'};
         return next(err);
     }
-    // if (spotStart) {
-    //     const err = new Error ('Sorry, this spot is already booked for the specifed dates');
-    //     err.status = 403;
-    //     err.errors = {startDate: 'Start date conflicts with an existing booking'};
-    //     return next(err);
-    // }
-    // if (spotEnd) {
-    //     const err = new Error ('Sorry, this spot is already booked for the specifed dates');
-    //     err.status = 403;
-    //     err.errors = {endDate: 'End date conflicts with an existing booking'};
-    //     return next(err);
-    // }
     return next()
-}
-
-
-
-
-
-
-const pastStartDate = async (req, res, next) => {
-
-    const booking = await Booking.findByPk(req.params.bookingId, {
-        attributes: ['startDate']
-    });
-    const todayWithTime = new Date();
-
-    const today = new Date(todayWithTime.getFullYear(), todayWithTime.getMonth(), todayWithTime.getDate())
-
-    const bookingStartDate = new Date(booking.startDate)
-
-    // console.log('start date', bookingStartDate)
-    // console.log('today', today)
-
-    if (today.getTime() > bookingStartDate.getTime()) {
-        const err = new Error ("Bookings that have been started can't be changed or deleted");
-        err.status = 400;
-        return next(err);
-    }
-    return next();
 }
 
 
@@ -170,7 +79,7 @@ router.put('/:bookingId', existsBooking, requireAuth, bookingBelongsPermission, 
         startDate,
         endDate
     });
-    // console.log('editBooking', editBooking)
+
     return res.json(editBooking);
 });
