+++ backend/routes/api/images.js
@@ -2,8 +2,7 @@ const express = require('express');
 const router = express.Router();
 
 const { requireAuth } = require('../../utils/auth');
-const { Spot, User, Review, Image, sequelize } = require('../../db/models');
-
+const { Spot, User, Review, Image } = require('../../db/models');
 
 
 const existsImage = async (req, res, next) => {
@@ -56,38 +55,6 @@ const imagePermission = async (req, res, next) => {
 
     return next();
 }
-
-// const imagePer = async (req, res, next) => {
-//     const spot = await Spot.findOne({
-//         where: {
-//           ownerId: req.user.id
-//         },
-//         include: {
-//             model: Image,
-//             where: {
-//                 id: req.params.imageId
-//             }
-//         }
-//       })
-
-//       const review = await Review.findOne({
-//         where: {
-//           userId: req.user.id
-//         },
-//         include: {
-//             model: Image,
-//             where: {
-//                 id: req.params.imageId
-//             }
-//         }
-//       })
-//       if (!spot && !review) {
-//         const err = new Error ('Forbidden');
-//         err.status = 403;
-//         return next(err);
-//       }
-//         return next();
-// }
 
 // Delete an Image
 router.delete('/:imageId', existsImage, requireAuth, imagePermission, async (req, res, next) => {
