+++ backend/routes/api/index.js
@@ -8,22 +8,6 @@ const reviewsRouter = require('./reviews.js');
 const yourBookingsRouter = require('./your-bookings.js');
 const bookingsRouter = require('./bookings.js');
 const imagesRouter = require('./images.js');
-// const { setTokenCookie } = require('../../utils/auth.js');
-// const { User } = require('../../db/models');
-
-// router.get('/set-token-cookie', async (_req, res) => {
-//   const user = await User.findOne({
-//       where: {
-//         username: 'Demo-lition'
-//       }
-//     });
-//   setTokenCookie(res, user);
-//   return res.json({ user });
-// });
-
-// router.post('/test', (req, res) => {
-//     res.json({requestBody: req.body});
-// });
 
 // GET /api/restore-user
 const { restoreUser } = require('../../utils/auth.js');
@@ -40,26 +24,5 @@ router.use('/reviews', reviewsRouter);
 router.use('/your-bookings', yourBookingsRouter);
 router.use('/bookings', bookingsRouter);
 router.use('/images', imagesRouter);
-
-// router.post('/test', (req, res) => {
-//     res.json({ requestBody: req.body });
-//   });
-
-// router.get(
-//   '/restore-user',
-//   (req, res) => {
-//     return res.json(req.user);
-//   }
-// );
-
-// GET /api/require-auth
-// const { requireAuth } = require('../../utils/auth.js');
-// router.get(
-//     '/require-auth',
-//     requireAuth,
-//     (req, res) => {
-//         return res.json(req.user);
-//     }
-// );
 
 module.exports = router;
